+++ source/dotnet/Measurements.Abstractions/Api/Models/MeasurementAggregationByDateDto.cs
@@ -9,4 +9,4 @@
 /// <param name="Unit">Unit of aggregated measurement.</param>
 /// <param name="MissingValues">Indicates whether aggregate contains missing values during day of aggregation.</param>
 /// <param name="ContainsUpdatedValues">Indicates whether any measurements in aggregate has been updated since creation.</param>
-public sealed record MeasurementAggregationDto(DateOnly Date, decimal Quantity, Quality Quality, Unit Unit, bool MissingValues, bool ContainsUpdatedValues);
+public sealed record MeasurementAggregationByDateDto(DateOnly Date, decimal Quantity, Quality Quality, Unit Unit, bool MissingValues, bool ContainsUpdatedValues);