+++ source/dotnet/Measurements.Abstractions/Api/Queries/GetAggregatedByMonthQuery.cs
@@ -3,6 +3,6 @@
 namespace Energinet.DataHub.Measurements.Abstractions.Api.Queries;
 
 /// <summary>
-/// Query for getting aggregated measurements for a metering point in a month specified by year and month.
+/// Query for getting measurements aggregated by day for a metering point specified by year and month.
 /// </summary>
 public sealed record GetAggregatedByMonthQuery(string MeteringPointId, YearMonth YearMonth);