+++ source/dotnet/Measurements.Infrastructure/Handlers/MeasurementsHandler.cs
@@ -3,6 +3,7 @@ using Energinet.DataHub.Measurements.Application.Handlers;
 using Energinet.DataHub.Measurements.Application.Persistence;
 using Energinet.DataHub.Measurements.Application.Requests;
 using Energinet.DataHub.Measurements.Application.Responses;
+using Energinet.DataHub.Measurements.Domain;
 using NodaTime;
 using NodaTime.Extensions;
 
@@ -29,13 +30,23 @@ public class MeasurementsHandler(IMeasurementsRepository measurementsRepository)
         return GetMeasurementResponse.Create(foundMeasurements);
     }
 
-    public async Task<GetAggregatedMeasurementsResponse> GetAggregatedByMonthAsync(GetAggregatedByMonthRequest request)
+    public async Task<GetMeasurementsAggregatedByDateResponse> GetAggregatedByDateAsync(GetAggregatedByDateRequest request)
     {
         var yearMonth = new YearMonth(request.Year, request.Month);
         var aggregatedMeasurements = await measurementsRepository
-            .GetAggregatedByMonthAsync(request.MeteringPointId, yearMonth)
+            .GetAggregatedByDateAsync(request.MeteringPointId, yearMonth)
             .ToListAsync() ?? throw new MeasurementsNotFoundDuringPeriodException();
 
-        return GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements);
+        return GetMeasurementsAggregatedByDateResponse.Create(aggregatedMeasurements);
+    }
+
+    public async Task<GetMeasurementsAggregatedByMonthResponse> GetAggregatedByMonthAsync(GetAggregatedByMonthRequest request)
+    {
+        var year = new Year(request.Year);
+        var aggregatedMeasurements = await measurementsRepository
+            .GetAggregatedByMonthAsync(request.MeteringPointId, year)
+            .ToListAsync() ?? throw new MeasurementsNotFoundDuringPeriodException();
+
+        return GetMeasurementsAggregatedByMonthResponse.Create(aggregatedMeasurements);
     }
 }